Moved Unit-related functions from OGFILE3 to new file OGFILE_UNITS.cpp.
